long long -> int

+++ HW6/Film.h
@@ -2,14 +2,14 @@
 #define FILM_H
 
 struct Film {
-  long long userId;
-  long long itemId;
+  int userId;
+  int itemId;
   double rating;
 
   Film() {
   }
   
-  Film(long long userId, long long itemId, double rating) : userId(userId), itemId(itemId), rating(rating) {
+  Film(int userId, int itemId, double rating) : userId(userId), itemId(itemId), rating(rating) {
   }
 };
